Update files for product management and generate PDF
Files regarding product management were updated. This includes data on product removal, editing, and registration. In addition, the product repository was created and PHP web settings were updated to include paths for third-party libraries. This was all part of a larger change to introduce a feature for generating PDFs of product data lists.